Fixed an issue

+++ modules/med/controllers.js
@@ -13,7 +13,7 @@ const addMed = async (req, res) => {
     return res.status(httpStatus.OK).json(newMed);
   } catch (error) {
     console.log('[AddMed] er: ', error);
-    await medSchema.findByIdAndDelete(newMed._id)
+    if (newMed && newMed._id) await medSchema.findByIdAndDelete(newMed._id)
     return res.status(httpStatus.BAD_REQUEST).json(error);
   }
 }
